ENH: Add warning if target is more specific than partitioning_version
The partitioning versions which were provided with the BIOSCAN-1M
dataset are only designed to work with the order level or family
level taxonomic ranks. Trying to use them with more detailed
taxonomic levels will not work well since there will be genera
(genuses) which appear in the test partition which aren't in the
train partition, making it infeasible for a supervised model to
learn to classify them.

We add a warning to alert a user when there is a mismatch between
the partition version they are using and the target type they
have requested.