MNT: Vectorize image_path code
Using `df.apply` to process each row individually in order to
generate the image_path column was taking 4 seconds for the train
partition (10% of the total class instantiation time).
This compute time scales linearly with the number of samples, so
when using all partitions this was taking 45-50 seconds
(55-75% of the total class instantiation time).

By vectorizing this code, it now only takes 1.6 seconds to add the
image_path column even when using all partitions.