ENH: Add clibd_split column when loading all data in CLIBD mode
If the partitioning_version is set to "clibd", and no split was
selected, we add a `clibd_split` column to show the CLIBD split.

This isn't added if the partitioning version is set to something
else since the CLIBD partition files can't be assumed to be present.

This isn't added if the split value was set, since we don't know
the actual split value if we load a CLIBD metasplit of
"all_keys" or "no_split_and_seen_train", and it would be cumbersome
to also load the appropriate files to figure it out.