BUG: Check images present by split before download and extract
Rather than downloading every split which is present in the
metasplit if any are missing, we go through each split and check
for its files before deciding to download and extract its zip.
This avoids re-downloading large amounts of data if (for example)
pretrain was already downloaded, then the zip file was deleted, and
now the user requests the "all" or "pretrain+train" metasplits.