MNT: Include processid column in BS1M metadata attribute
The processid is the record label used by BOLD and should over the
sampleid. This is used by BIOSCAN5M for the image paths, so adding
it to always be in BIOSCAN1M's metadata allows for aligning the
records between datasets more easily and clearly.

We retain the sampleid column to use for selecting the CLIBD
splits, and for backward compatibility since this column was
loaded previously. The CLIBD partitions are done by sampleid
since this is in the CLIBD HDF5 files but processid is (last I
checked) not present.