BUG: Fix handling of BIOSCAN5M(..., split=None)
This was not indicated as supported in the docstring, but was
in the type hint and did not work because explode_metasplit expects
a string. Using split=None was not causing an issue for
BIOSCAN1M, load_bioscan1m_metadata, or load_bioscan5m_metadata,
only BIOSCAN5M was affected.